Backend 2/requirements.txt
fastapi
uvicorn
pandas
openpyxl
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
sqlalchemy
psycopg2-binary
rapidfuzz
python-multipart
